refactor(components): [cascader] use script-setup syntax
